move code to enterprise_gwecc and update version
