--- examples/pdist_prior.py
@@ -1,1 +1,0 @@
-../src/pdist_prior.py
