+++ setup.py
@@ -2,12 +2,12 @@ from setuptools import setup
 
 setup(
     name="enterprise_gwecc",
-    version="0.1.0",
-    description="Computes pulsar TOA delays due to gravitational waves from eccentric supermassive binary sources.",
+    version="0.1.1",
+    description="Compute pulsar timing array signals due to eccentric supermassive binaries",
     author="Abhimanyu Susobhanan",
     author_email="abhisrkckl@gmail.com",
     python_requires=">=3.8",
     package_dir={"": "src"},
     py_modules=["enterprise_gwecc"],
-    install_requires=["numpy", "matplotlib", "enterprise-pulsar", "juliacall"],
+    install_requires=["numpy", "scipy", "matplotlib", "enterprise-pulsar", "juliacall"],
 )
